requirements.txt
tqdm==4.19.9
pytorch==0.3.1
pytz==2017.3
requests==2.18.4
numpy==1.13.3
scipy==1.0.0
Pillow==5.1.0
visdom

